+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - (Feature) PongV1 Integration Service
 - (Feature) Custom Gateway image
 - (Bugfix) Fix race condition in ArangoBackup
+- (Feature) Improve Gateway Config gen
 
 ## [1.2.42](https://github.com/arangodb/kube-arangodb/tree/1.2.42) (2024-07-23)
 - (Maintenance) Go 1.22.4 & Kubernetes 1.29.6 libraries